\name{ibs-package}
\alias{ibs-package}
\docType{package}
\title{
  Integrated B-spline function
}
\description{
  Evaluate the integral of a B-spline function, or the B-spline function
  itself, or the B-spline basis functions.
}
\details{
\tabular{ll}{
Package: \tab ibs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-04-12\cr
License: \tab GPL (>=2)\cr
}
}
\author{
Feng Chen <feng.chen@unsw.edu.au>
Maintainer: Feng Chen <feng.chen@unsw.edu.au>
}
\references{
  de Boor, C. (2001) A Practical Guide to Splines. Revised Edition. Springer: New York.
}
\keyword{ package }
\seealso{
  \code{\link[splines]{splineDesign}}
}

